# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: Futsal Reservation API
  description: |
    API complète pour l'application de réservation de matchs de futsal.
    
    Cette API permet de gérer :
    - L'authentification des utilisateurs
    - La gestion des matchs (création, inscription, désinscription)
    - Les notifications et annonces
    - L'administration des utilisateurs et statistiques
    
    ## Authentification
    L'API utilise Better-auth avec des sessions sécurisées. Les cookies sont automatiquement gérés par le navigateur.
    
    ## Rate Limiting
    - Login : 5 tentatives/minute par IP
    - Register : 3 tentatives/minute par IP
    - API générale : Protection par utilisateur authentifié
    
    ## Codes d'erreur
    - 400: Requête invalide (validation échouée)
    - 401: Non authentifié
    - 403: Accès refusé (permissions insuffisantes)
    - 404: Ressource non trouvée
    - 429: Limite de débit dépassée
    - 500: Erreur serveur interne
    
  version: 1.0.0
  contact:
    name: Futsal App Support
    email: support@futsal.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Serveur de développement
  - url: https://api.futsal.app
    description: Serveur de production
...
paths:
...
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Connexion utilisateur
      description: Authentifie un utilisateur et démarre une session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Identifiants incorrects
        '429':
          description: Trop de tentatives
  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: Déconnexion utilisateur
      description: Termine la session utilisateur
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Déconnexion réussie
...
  /api/admin/users:
    get:
      tags:
        - Admin
      summary: Liste des utilisateurs (Admin)
      security:
        - cookieAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Liste des utilisateurs
        '403':
          description: Accès refusé
  /api/admin/statistics:
    get:
      tags:
        - Admin
      summary: Statistiques globales (Admin)
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Statistiques
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AdminStatistics'

  # Health endpoint
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
  schemas:
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        pseudo:
          type: string
        avatar:
          type: string
          format: uri
        role:
          type: string
          enum: [user, root]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AuthResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                user:
                  $ref: '#/components/schemas/User'
...
    AdminStatistics:
      type: object
      properties:
        users:
          type: object
          properties:
            total:
              type: integer
            active:
              type: integer
            newThisMonth:
              type: integer
        matches:
          type: object
          properties:
            total:
              type: integer
            thisMonth:
              type: integer
            upcoming:
              type: integer
            averageParticipants:
              type: number
...